requirements.txt
flask
flask-cors
python-dotenv
pymongo
pandas
deepface
requests
gunicorn
